Added item package and fixed issues in inventory class.
